app/requirements.txt
schedule
Flask
pytube
beautifulsoup4
lxml
pygame
django
pyqt5
django-datetime-widget